backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic-settings==2.4.0
httpx==0.27.0
beautifulsoup4==4.12.3
python-multipart==0.0.9
pillow==10.4.0
pandas==2.2.2
numpy==2.1.1
docling==2.8.3
llama-cpp-python==0.3.16
huggingface_hub==0.24.5
transformers>=4.38.0
torch>=2.1.0
torchvision>=0.16.0
python-doctr[torch]>=0.8.0
timm>=1.0.0
pytest==8.3.3
